Change poll times to doubles in constructor, but still int in C linkage for vxWorks

+++ motorApp/ACRSrc/ACRMotorDriver.h
@@ -50,7 +50,7 @@ friend class ACRController;
 
 class ACRController : public asynMotorController {
 public:
-  ACRController(const char *portName, const char *ACRPortName, int numAxes, int movingPollPeriod, int idlePollPeriod);
+  ACRController(const char *portName, const char *ACRPortName, int numAxes, double movingPollPeriod, double idlePollPeriod);
   
   /* These are the methods that we override from asynMotorDriver */
   asynStatus writeInt32(asynUser *pasynUser, epicsInt32 value);
